+++ packages/astro/src/core/app/node.ts
@@ -1,11 +1,12 @@
 import type { RouteData } from '../../@types/astro.js';
+import type { RenderOptions } from './index.js';
 import type { SerializedSSRManifest, SSRManifest } from './types.js';
 
 import * as fs from 'node:fs';
 import { IncomingMessage } from 'node:http';
 import { TLSSocket } from 'node:tls';
 import { deserializeManifest } from './common.js';
-import { App, type MatchOptions } from './index.js';
+import { App } from './index.js';
 export { apply as applyPolyfills } from '../polyfill.js';
 
 const clientAddressSymbol = Symbol.for('astro.clientAddress');
@@ -108,19 +109,34 @@ class NodeIncomingMessage extends IncomingMessage {
 }
 
 export class NodeApp extends App {
-	match(req: NodeIncomingMessage | Request, opts: MatchOptions = {}) {
+	match(req: NodeIncomingMessage | Request) {
 		if (!(req instanceof Request)) {
 			req = createRequestFromNodeRequest(req, {
 				emptyBody: true,
 			});
 		}
-		return super.match(req, opts);
+		return super.match(req);
 	}
-	render(req: NodeIncomingMessage | Request, routeData?: RouteData, locals?: object) {
+	render(request: NodeIncomingMessage | Request, options?: RenderOptions): Promise<Response>;
+	/**
+	 * @deprecated Instead of passing `RouteData` and locals individually, pass an object with `routeData` and `locals` properties.
+	 * See https://github.com/withastro/astro/pull/9199 for more information.
+	 */
+	render(
+		request: NodeIncomingMessage | Request,
+		routeData?: RouteData,
+		locals?: object
+	): Promise<Response>;
+	render(
+		req: NodeIncomingMessage | Request,
+		routeDataOrOptions?: RouteData | RenderOptions,
+		maybeLocals?: object
+	) {
 		if (!(req instanceof Request)) {
 			req = createRequestFromNodeRequest(req);
 		}
-		return super.render(req, routeData, locals);
+		// @ts-expect-error The call would have succeeded against the implementation, but implementation signatures of overloads are not externally visible.
+		return super.render(req, routeDataOrOptions, maybeLocals);
 	}
 }