+++ packages/astro/src/i18n/vite-plugin-i18n.ts
@@ -1,4 +1,4 @@
-import * as vite from 'vite';
+import type * as vite from 'vite';
 import type { AstroSettings } from '../@types/astro.js';
 
 const virtualModuleId = 'astro:i18n';
@@ -29,7 +29,7 @@ export default function astroInternationalization({
 						getLocaleAbsoluteUrlList as _getLocaleAbsoluteUrlList,
 						getPathByLocale as _getPathByLocale, 
 						getLocaleByPath as _getLocaleByPath,
-					} from "astro/i18n";
+					} from "astro/virtual-modules/i18n.js";
 					
 					const base =  ${JSON.stringify(settings.config.base)};
 					const trailingSlash =  ${JSON.stringify(settings.config.trailingSlash)};