+++ packages/astro/src/runtime/server/render/common.ts
@@ -67,8 +67,8 @@ function stringifyChunk(
 				let prescriptType: PrescriptType = needsHydrationScript
 					? 'both'
 					: needsDirectiveScript
-					? 'directive'
-					: null;
+					  ? 'directive'
+					  : null;
 				if (prescriptType) {
 					let prescripts = getPrescripts(result, prescriptType, hydration.directive);
 					return markHTMLString(prescripts);