+++ packages/astro/test/fixtures/core-image-ssr/src/pages/api.ts
@@ -3,8 +3,5 @@ import type { APIRoute } from "../../../../../src/@types/astro";
 export const GET = (async ({ params, request }) => {
 	const url = new URL(request.url);
   const src = url.searchParams.get("src");
-
-	return {
-		body: "An image: " + JSON.stringify(src),
-	};
+	return new Response("An image: " + JSON.stringify(src));
 }) satisfies APIRoute;