+++ packages/integrations/mdx/test/fixtures/mdx-frontmatter-injection/src/pages/glob.json.js
@@ -1,6 +1,6 @@
-export async function get() {
+export async function GET() {
 	const docs = await import.meta.glob('./*.mdx', { eager: true });
-	return {
-		body: JSON.stringify(Object.values(docs).map(doc => doc.frontmatter)),
-	}
+	return new Response(
+		JSON.stringify(Object.values(docs).map(doc => doc.frontmatter))
+	);
 }